Add Najman, Couprie maxtree algorithm + fix algorithm to use image size_type instead of std::size_t.
	* mln/core/algorithm/sort_indexes.hpp,
	* mln/core/image/image2d.hpp,
	* mln/core/image/ndimage.hpp,
	* mln/morpho/bounded_hqueue.hpp,
	* mln/morpho/canonize.hpp,
	* mln/morpho/maxtree_hqueue.hpp,
	* mln/morpho/maxtree_hqueue_parallel.hpp,
	* mln/morpho/maxtree_pqueue.hpp,
	* mln/morpho/maxtree_pqueue_parallel.hpp,
	* mln/morpho/maxtree_routines.hpp,
	* mln/morpho/maxtree_ufind_parallel.hpp,
	* mln/morpho/maxtree_ufindrank_parallel.hpp,
	* mln/morpho/merge_tree.hpp,
	* mln/morpho/pqueue_fast.hpp,
	* tests/core/algorithm/sort_indexe.cpp,
	* tests/morpho/bench_maxtree.cpp,
	* tests/morpho/bench_maxtree_algorithm.hpp,
	* tests/morpho/maxtree.cpp: Fix to use Image::size_type instead of size_t.
	* mln/morpho/maxtree_najman.hpp,
	* tests/morpho/bench_serial_najman.cpp,
	* tests/morpho/CMakeLists.txt: New. Najman, Couprie Implementation.

+++ mln/core/image/ndimage.hpp
@@ -86,8 +86,8 @@ namespace mln
     typedef T*			pointer;
     typedef const T*		const_pointer;
     typedef ptrdiff_t		difference_type;
-    typedef size_t		index_type;
-    typedef size_t		size_type;
+    typedef unsigned		index_type;
+    typedef unsigned		size_type;
